requirements_dev.txt
currencyconverter==0.17.7
pytest==7.3.2
apache-airflow==2.6.1
pandas==2.0.2